refactor: move error_tracking_adapter from _core/ to _persistence/
The adapter bridges _core.ErrorTracker and _persistence.RecoveryManager
into CacheRecoveryPort. Placing it in _core/ created an upward dependency
from the foundation layer into _persistence/. Moving it to _persistence/
eliminates this Clean Architecture violation.

The old location (_core/error_tracking_adapter.py) retains a re-export
shim for backward compatibility.
